requirements.txt
flask
flask_restful
web3
python-dotenv
pydantic
flask_pydantic
flask_cors
flask_login
py-solc-x
cmake
pyjwt
retrying
